packages/scraper/pyproject.toml
[project]
name = "parseltongue-scraper"
version = "0.1.0"
description = "Phase 1: Scrape fan fiction (AO3) to Markdown + metadata"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
